ci: Add Markdown Link Checks (#48)
# Pull Request

## Description

This pull request introduces a new workflow to check for broken Markdown
links in the repository. The changes include adding a new job to the
GitHub Actions workflow and configuring the link checker.

### New workflow for checking Markdown links:

*
[`.github/other-configurations/.linkspector.yml`](diffhunk://#diff-f9691f23ea6c4c34ecbaa23a98721e0aa3c7a94f2fcdd38f7b99fe863c5f6305R1-R5):
Added configuration file for the link checker specifying directories to
scan and valid status codes.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R46-R63):
Added a new job `check-markdown-links` to the GitHub Actions workflow to
automate the process of checking Markdown links using
`UmbrellaDocs/action-linkspector`.

Fixes #12