chore: Add Code of Conduct (#47)
# Pull Request

## Description

This pull request introduces a new Code of Conduct document to the
project. The document outlines the expected behaviors and
responsibilities of community members, as well as the procedures for
enforcing the code.

Key changes include:

* **Introduction of Code of Conduct**:
* Added a comprehensive Code of Conduct based on the Contributor
Covenant, version 2.1, to ensure a harassment-free experience for all
community members.

* **Standards and Examples**:
* Defined acceptable and unacceptable behaviors to foster a positive and
inclusive environment.

* **Enforcement Responsibilities**:
* Clarified the responsibilities of community leaders in enforcing the
Code of Conduct and taking corrective actions.

* **Scope and Enforcement Guidelines**:
* Specified the scope of the Code of Conduct and detailed the
enforcement guidelines, including potential consequences for violations.

* **Attribution**:
* Provided attribution to the original Contributor Covenant and
Mozilla's code of conduct enforcement ladder, along with links to
additional resources and translations.

Fixes #4