chore: Decomission Project (#77)
# Pull Request

## Description

This pull request includes a small but important change to the
`README.md` file. The change adds a warning to inform users that the
project has been terminated and is no longer under development or
maintenance.

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R10-R12):
Added a warning section to indicate that the project is no longer under
development/maintenance.